Adds improved visual aspects and reduces memory footprint
This commit adds animations for the player, as well as improved sprites
for walls and floors. It also implements wall shadows, making the game
have a more isometric feel, enables pooling in PathfindingGrid.Java, fixes
texture bleeding, adds debug-drawing for the collidable tiles and cleans
up a big portion of the code base.