Implement enemies, barriers and a better combat structure
This commit will add randomly generate spawn points for which
the enemies can breach through (though after having them torn
down). It will implement better animations for the axe, more
precise hitboxes and a more sufficient enemy AI. The spritesheet
will be more organised and minor changes as the origin of the
entities from the left-bottom corner change to the centre and
helper methods for length between instances within the map
will also be implemented.