Add traps, bombs, improved AI and a wave system
This commit will add step traps which can be crafted with potions
and wood and once stepped on, they'll break the potions, resulting
in an explosion or just liquid flowing out. Thrown potions with
a set flammability will explode. The enemies will now follow find
paths and follow the player through this path. This commit implement
this feature, though is it not perfect yet. This commit will also
implement a wave system which deploys n amount of enemies given
and arithmetic sequence and once all the enemies have died the
wave system counts down a given time period before the a new wave
is deployed.